WilsonMG: Move tests for Wilson & WilsonClover into the same file
